perbaiki fitur

+++ database/migrations/2022_05_24_073927_create_data_pemasukans_table.php
@@ -18,7 +18,7 @@ class CreateDataPemasukansTable extends Migration
             $table->foreignId('id_akun');
             $table->string('nama_pemasukan', 30);
             $table->smallInteger('jumlah');
-            $table->float('harga_satuan');
+            $table->double('harga_satuan');
             $table->string('keterangan', 50);
             $table->timestamps();
