+++ database/migrations/2022_05_24_074703_create_data_pengeluarans_table.php
@@ -18,7 +18,7 @@ class CreateDataPengeluaransTable extends Migration
             $table->foreignId('id_akun');
             $table->string('nama_pengeluaran', 30);
             $table->smallInteger('jumlah');
-            $table->float('harga_satuan');
+            $table->double('harga_satuan');
             $table->string('keterangan', 50);
             $table->timestamps();
