+++ database/migrations/2022_05_24_074853_create_data_hutangs_table.php
@@ -18,7 +18,7 @@ class CreateDataHutangsTable extends Migration
             $table->foreignId('id_akun');
             $table->string('nama_hutang', 30);
             $table->smallInteger('jumlah');
-            $table->float('harga_satuan');
+            $table->double('harga_satuan');
             $table->string('keterangan', 50);
             $table->timestamps();
